pyproject.toml
[project]
name = "courses"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "markdown-it-py>=3.0.0",
    "mkslides>=1.0.10",
    "playwright>=1.55.0",
]
